SAK-41175 Allow floating point numbers in rubrics points (#7853)
https://jira.sakaiproject.org/browse/SAK-41175
(cherry picked from commit 3218c840b73b0ae0308814c050c4b68034a3ea7a)

+++ rubrics/api/src/main/java/org/sakaiproject/rubrics/logic/model/CriterionOutcome.java
@@ -67,7 +67,7 @@ public class CriterionOutcome implements Serializable {
         private boolean pointsAdjusted;
 
         @NonNull
-        private Integer points;
+        private Double points;
 
         @Lob
         @Column(length = 65535)
